perf: align multiply/normalize/weighted paths with 0.0.40 router baseline
- multiply_by_constant float32: multiply_numpy (0.0.40), not OpenCV/NumKong
- add_weighted float32: cv2.addWeighted; uint8 keeps nk.blend
- normalize: global 3D stats via img.mean/img.std in opencv helper; numpy 'image' branch same
- pairwise: keep nk.cdist for n1*n2<1000 (numpy-only was slower here)
- benchmark: uint8 channel vectors as uint8 for add_vector/multiply_by_vector
- Refresh router JSON + REPORT vs frozen 0.0.40; update numkong + regression docs

Made-with: Cursor

+++ benchmarks/benchmark_router_synthetic.py
@@ -126,7 +126,7 @@ def _bench(
 def _channel_vector(img: np.ndarray) -> np.ndarray:
     c = img.shape[-1]
     if img.dtype == np.uint8:
-        return np.full(c, 2, dtype=np.int32)
+        return np.full(c, 2, dtype=np.uint8)
     return np.linspace(0.98, 1.02, num=c, dtype=np.float32)
